URL, 로직 수정

See merge request minjikim99/devroutine!28

+++ src/main/java/com/likelion/devroutine/participant/service/ParticipationService.java
@@ -5,9 +5,17 @@ import com.likelion.devroutine.challenge.dto.ChallengeDto;
 import com.likelion.devroutine.challenge.exception.ChallengeNotFoundException;
 import com.likelion.devroutine.challenge.exception.InProgressingChallengeException;
 import com.likelion.devroutine.challenge.exception.InaccessibleChallengeException;
+import com.likelion.devroutine.challenge.exception.InvalidPermissionException;
+import com.likelion.devroutine.follow.domain.Follow;
+import com.likelion.devroutine.follow.dto.FollowerResponse;
+import com.likelion.devroutine.follow.exception.FollowingNotFoundException;
+import com.likelion.devroutine.follow.repository.FollowRepository;
 import com.likelion.devroutine.hashtag.domain.ChallengeHashTag;
 import com.likelion.devroutine.hashtag.dto.ChallengeHashTagResponse;
 import com.likelion.devroutine.hashtag.repository.ChallengeHashTagRepository;
+import com.likelion.devroutine.invite.domain.Invite;
+import com.likelion.devroutine.invite.dto.InviteCreateResponse;
+import com.likelion.devroutine.invite.repository.InviteRepository;
 import com.likelion.devroutine.participant.domain.Participation;
 import com.likelion.devroutine.participant.dto.ParticipationChallengeDto;
 import com.likelion.devroutine.participant.dto.ParticipationResponse;
@@ -15,6 +23,7 @@ import com.likelion.devroutine.participant.enumerate.ResponseMessage;
 import com.likelion.devroutine.participant.exception.DuplicatedParticipationException;
 import com.likelion.devroutine.challenge.repository.ChallengeRepository;
 import com.likelion.devroutine.participant.exception.ParticipationNotFoundException;
+import com.likelion.devroutine.participant.exception.RejectCancelException;
 import com.likelion.devroutine.participant.repository.ParticipationRepository;
 import com.likelion.devroutine.user.domain.User;
 import com.likelion.devroutine.user.dto.UserResponse;
@@ -34,30 +43,26 @@ public class ParticipationService {
     private final ParticipationRepository participationRepository;
     private final UserRepository userRepository;
     private final ChallengeRepository challengeRepository;
+    private final InviteRepository inviteRepository;
+    private final FollowRepository followRepository;
     private final ChallengeHashTagRepository challengeHashTagRepository;
 
-    public ParticipationService(ParticipationRepository participationRepository, UserRepository userRepository, ChallengeRepository challengeRepository, ChallengeHashTagRepository challengeHashTagRepository) {
+    public ParticipationService(ParticipationRepository participationRepository, UserRepository userRepository, ChallengeRepository challengeRepository, InviteRepository inviteRepository, FollowRepository followRepository) {
         this.participationRepository = participationRepository;
         this.userRepository = userRepository;
         this.challengeRepository = challengeRepository;
+        this.inviteRepository = inviteRepository;
+        this.followRepository = followRepository;
         this.challengeHashTagRepository = challengeHashTagRepository;
-    }
-    @Transactional
-    public ParticipationResponse participateChallenge(String oauthId, Long challengeId) {
-        User user=getUser(oauthId);
-        Challenge challenge=getChallenge(challengeId);
-        validateParticipate(user, challenge);
-        Participation savedParticipation = participationRepository.save(Participation.createParticipant(user, challenge));
-        return ParticipationResponse.builder()
-                .challengeId(savedParticipation.getChallenge().getId())
-                .message(ResponseMessage.PARTICIPATE_SUCCESS.getMessage())
-                .build();
     }
 
     @Transactional
     public ParticipationResponse cancelChallenge(String oauthId, Long challengeId) {
         User user=getUser(oauthId);
         Challenge challenge=getChallenge(challengeId);
+        if(challenge.getUserId().equals(user.getId())){
+            throw new RejectCancelException();
+        }
         Participation participation =getParticipant(user, challenge);
         isProgressChallenge(challenge.getStartDate());
         participationRepository.delete(participation);
@@ -66,6 +71,33 @@ public class ParticipationService {
                 .message(ResponseMessage.CHALLENGE_CANCEL_SUCCESS.getMessage())
                 .build();
     }
+
+    @Transactional
+    public InviteCreateResponse inviteUser(String inviterOauthId, Long challengeId, Long inviteeId) {
+        User inviter=getUser(inviterOauthId);
+        User invitee=userRepository.findById(inviteeId)
+                .orElseThrow(()->new UserNotFoundException());
+        validateFollower(inviter, invitee);
+        Challenge challenge=getChallenge(challengeId);
+        matchWriterAndUser(inviter, challenge);
+        isProgressChallenge(challenge.getStartDate());
+        Invite savedInvite=inviteRepository.save(Invite.createInvite(challenge.getId(), inviter.getId(), invitee.getId()));
+        return InviteCreateResponse.builder()
+                .inviterId(savedInvite.getInviterId())
+                .challengeId(savedInvite.getChallengeId())
+                .inviteeId(savedInvite.getInviteeId())
+                .message(com.likelion.devroutine.invite.enumerate.ResponseMessage.INVITE_SUCCESS.getMessage())
+                .build();
+    }
+
+    public List<FollowerResponse> findFollowers(String oauthId, Long challengeId) {
+        User user = getUser(oauthId);
+        Challenge challenge=getChallenge(challengeId);
+        matchWriterAndUser(user, challenge);
+        List<Follow> followers=followRepository.findByFollowerId(user.getId());
+        return FollowerResponse.of(followers);
+    }
+
     public ParticipationChallengeDto findByParticipateChallenge(String oauthId, Long challengeId){
         User user=getUser(oauthId);
         Challenge challenge=getChallenge(challengeId);
@@ -73,10 +105,12 @@ public class ParticipationService {
         List<Participation> participations = participationRepository.findAllByChallenge(challenge);
         return ParticipationChallengeDto.toResponse(participation, ChallengeHashTagResponse.of(challenge.getChallengeHashTags()), UserResponse.toList(getChallengeParticipants(participations)));
     }
+
     public List<User> getChallengeParticipants(List<Participation> participations){
         return participations.stream().map(participant -> participant.getUser())
                 .collect(Collectors.toList());
     }
+
     public Participation getParticipant(User user, Challenge challenge){
         return participationRepository.findByUserAndChallenge(user, challenge)
                 .orElseThrow(()-> new ParticipationNotFoundException());
@@ -94,24 +128,11 @@ public class ParticipationService {
         return challenge;
     }
 
-    public boolean validateDuplicateParticipate(User user, Challenge challenge){
-        participationRepository.findByUserAndChallenge(user, challenge)
-                .ifPresent(participant->{
-                    throw new DuplicatedParticipationException();
-                });
-        return true;
-    }
-
-    public boolean validateParticipate(User user, Challenge challenge){
-        isVigibility(challenge);
-        isProgressChallenge(challenge.getStartDate());
-        validateDuplicateParticipate(user, challenge);
-        return true;
-    }
-
-    public boolean isVigibility(Challenge challenge) {
-        if (challenge.getVigibility()) return true;
-        else throw new InaccessibleChallengeException();
+    public boolean isViewable(Challenge challenge, User user){
+        if(challenge.getVigibility() || !inviteRepository.findAllByChallengeIdAndInviteeId(challenge.getId(), user.getId()).isEmpty()){
+            return true;
+        }
+        throw new InaccessibleChallengeException();
     }
 
     public boolean isProgressChallenge(LocalDate startDate) {
@@ -119,6 +140,17 @@ public class ParticipationService {
             throw new InProgressingChallengeException();
         return false;
     }
+
+    private void validateFollower(User inviter, User invitee) {
+        Follow follow=followRepository.findByFollowerIdAndFollowingId(inviter.getId(), invitee.getId())
+                .orElseThrow(()->new FollowingNotFoundException());
+    }
+
+    private boolean matchWriterAndUser(User user, Challenge challenge) {
+        if(!user.getId().equals(challenge.getUserId())){
+            throw new InvalidPermissionException();
+        }
+        return true;
 
     public List<ChallengeDto> findAllParticipateChallenge(String oauthId){
         User user=getUser(oauthId);
@@ -129,6 +161,7 @@ public class ParticipationService {
                 .collect(Collectors.toList());
         return ChallengeDto.toList(participateChallenges, getChallengeHashTagResponse(participateChallenges));
     }
+
     private Map<Long, List<ChallengeHashTagResponse>> getChallengeHashTagResponse(List<Challenge> challenges){
         return challenges
                 .stream()
@@ -137,6 +170,7 @@ public class ParticipationService {
                         challenge->ChallengeHashTagResponse.of(getHashTags(challenge.getId()))
                 ));
     }
+
     private List<ChallengeHashTag> getHashTags(Long challengeId) {
         return challengeHashTagRepository.findByChallengeId(challengeId);
     }
